+++ Sources/Utils/CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.38 (pre-release):
 * [Enhancement] Add a type formatter for the part argument: [`GUIUtils.PartType`](http://ihsoft.github.io/KSPDev/Utils/html/T_KSPDev_GUIUtils_PartType.htm).
 * [Enhancement] Add a type formatter for the fixed percent form: [`GUIUtils.PercentFixedType`](http://ihsoft.github.io/KSPDev/Utils/html/T_KSPDev_GUIUtils_PercentFixedType.htm).
+* [Enhancement] Introduce "Hermetic GUI controls" concept: [`GUIUtils.AbstractHermeticGUIControl`](http://ihsoft.github.io/KSPDev/Utils/html/T_KSPDev_GUIUtils_AbstractHermeticGUIControl.htm).
 
 # 0.37.0 (July 8th, 2018):
 * [Change] Don't reload the `KSPField` fields on language change to improve stability. Delegate it to the `Localization Tool`.